add ability to serve getblockheader requests + some cleanup
